backend/requirements.txt
Flask==3.1.1
Flask-Cors==5.0.0
python-dotenv==1.1.1
requests==2.32.4